+++ cinelerra/edlsession.h
@@ -99,8 +99,8 @@ public:
 // Coords for cropping operation
 	int crop_x1, crop_x2, crop_y1, crop_y2;
 // Ruler points relative to the output frame.
-	float ruler_x1, ruler_y1;
-	float ruler_x2, ruler_y2;
+	double ruler_x1, ruler_y1;
+	double ruler_x2, ruler_y2;
 // Current folder in resource window
 	int awindow_folder;
 // align cursor on frame boundaries
