+++ cinelerra/vdevicex11.h
@@ -145,8 +145,8 @@ private:
 // for last frame rendered.
 // These stick the last frame to the display.
 // Must be floats to support OpenGL
-	float output_x1, output_y1, output_x2, output_y2;
-	float canvas_x1, canvas_y1, canvas_x2, canvas_y2;
+	double output_x1, output_y1, output_x2, output_y2;
+	double canvas_x1, canvas_y1, canvas_x2, canvas_y2;
 // Set when OpenGL rendering has cleared the frame buffer before write_buffer
 	int is_cleared;
 // Set when OpenGL initialization failed
